+++ backend/lib/helpers/__init__.py
@@ -1,9 +1,9 @@
 from . import cache, checkers, commands
 from . import events, exceptions, locking
-from . import thread_actions, types
+from . import thread_actions
 
 __all__ = (
     'cache', 'checkers', 'commands',
     'events', 'exceptions', 'locking',
-    'thread_actions', 'types',
+    'thread_actions'
 )
