+++ backend/lib/models/__init__.py
@@ -4,10 +4,11 @@ from .game_state import GameState
 from .global_config import GlobalConfig
 from .task import Task
 from .team import Team
+from .types import Action, TaskStatus
 from .verdict import CheckerVerdict
 
 __all__ = (
     'AttackResult', 'Flag', 'GameState',
     'GlobalConfig', 'Task', 'Team',
-    'CheckerVerdict',
+    'Action', 'TaskStatus', 'CheckerVerdict',
 )
